split out top level parts of the ingest protocol into a standalone crate
