+++ client-libraries/rust/modality-ingest-client/Cargo.toml
@@ -10,6 +10,7 @@ homepage = "https://auxon.io/"
 [dependencies]
 minicbor = { version = "0.13.1", features = ["std", "derive"] }
 modality-api = { version = "0.1", path = "../modality-api" }
+modality-ingest-protocol = { version = "0.1", path = "../modality-ingest-protocol" }
 native-tls = "0.2.10"
 ordered-float = "2.2.0"
 thiserror = "1"
