+++ client-libraries/rust/modality-ingest-protocol/src/lib.rs
@@ -1,0 +1,3 @@
+mod protocol;
+
+pub use protocol::*;
